+++ Recyclable.Collections.Benchmarks/Benchmarks/Add_WithCapacity.cs
@@ -8,7 +8,7 @@ namespace Recyclable.Collections.Benchmarks
 		{
 			var data = TestObjects;
 			var dataCount = TestObjectCount;
-			var list = new List<long>(dataCount);
+			var list = new List<int>(dataCount);
 			for (var i = 0; i < dataCount; i++)
 			{
 				list.Add(data[i]);
@@ -19,7 +19,7 @@ namespace Recyclable.Collections.Benchmarks
 		{
 			var data = TestObjects;
 			var dataCount = TestObjectCount;
-			using var list = new PooledList<long>(capacity: dataCount, ClearMode.Auto);
+			using var list = new PooledList<int>(capacity: dataCount, ClearMode.Auto);
 			for (var i = 0; i < dataCount; i++)
 			{
 				list.Add(data[i]);
@@ -30,7 +30,7 @@ namespace Recyclable.Collections.Benchmarks
 		{
 			var data = TestObjects;
 			var dataCount = TestObjectCount;
-			using var list = new RecyclableList<long>(dataCount);
+			using var list = new RecyclableList<int>(dataCount);
 			for (var i = 0; i < dataCount; i++)
 			{
 				list.Add(data[i]);
@@ -41,7 +41,7 @@ namespace Recyclable.Collections.Benchmarks
 		{
 			var data = TestObjects;
 			var dataCount = TestObjectCount;
-			using var list = new RecyclableLongList<long>(minBlockSize: BlockSize, initialCapacity: dataCount);
+			using var list = new RecyclableLongList<int>(minBlockSize: BlockSize, initialCapacity: dataCount);
 			for (var i = 0; i < dataCount; i++)
 			{
 				list.Add(data[i]);
