+++ Recyclable.Collections.Benchmarks/Benchmarks/RemoveAt_FirstItems.cs
@@ -7,7 +7,7 @@ namespace Recyclable.Collections.Benchmarks
 		public void List_RemoveAt_FirstItems()
 		{
 			var data = TestObjects;
-			var list = new List<long>(data);
+			var list = new List<int>(data);
 			var dataCount = TestObjectCount / 10 > 0 ? TestObjectCount / 10 : TestObjectCount;
 			for (var i = 0; i < dataCount; i++)
 			{
@@ -18,7 +18,7 @@ namespace Recyclable.Collections.Benchmarks
 		public void PooledList_RemoveAt_FirstItems()
 		{
 			var data = TestObjects;
-			using var list = new PooledList<long>(data, ClearMode.Auto);
+			using var list = new PooledList<int>(data, ClearMode.Auto);
 			var dataCount = TestObjectCount / 10 > 0 ? TestObjectCount / 10 : TestObjectCount;
 			for (var i = 0; i < dataCount; i++)
 			{
@@ -29,7 +29,7 @@ namespace Recyclable.Collections.Benchmarks
 		public void RecyclableList_RemoveAt_FirstItems()
 		{
 			var data = TestObjects;
-			using var list = new RecyclableList<long>(data);
+			using var list = new RecyclableList<int>(data);
 			var dataCount = TestObjectCount / 10 > 0 ? TestObjectCount / 10 : TestObjectCount;
 			for (var i = 0; i < dataCount; i++)
 			{
@@ -40,7 +40,7 @@ namespace Recyclable.Collections.Benchmarks
 		public void RecyclableLongList_RemoveAt_FirstItems()
 		{
 			var data = TestObjects;
-			using var list = new RecyclableLongList<long>(data, minBlockSize: BlockSize, initialCapacity: base.TestObjectCount);
+			using var list = new RecyclableLongList<int>(data, minBlockSize: BlockSize, initialCapacity: base.TestObjectCount);
 			var dataCount = TestObjectCount / 10 > 0 ? TestObjectCount / 10 : TestObjectCount;
 			for (var i = 0; i < dataCount; i++)
 			{
