+++ Recyclable.Collections.Benchmarks/Benchmarks/Remove_LastItems.cs
@@ -7,7 +7,7 @@ namespace Recyclable.Collections.Benchmarks
 		public void List_Remove_LastItems()
 		{
 			var data = TestObjects;
-			var list = new List<long>(data);
+			var list = new List<int>(data);
 			var dataCount = TestObjectCount / 10 > 0 ? TestObjectCount / 10 : TestObjectCount;
 			for (var i = data.Length - 1; i > data.Length - dataCount - 1; i--)
 			{
@@ -18,7 +18,7 @@ namespace Recyclable.Collections.Benchmarks
 		public void PooledList_Remove_LastItems()
 		{
 			var data = TestObjects;
-			using var list = new PooledList<long>(data, ClearMode.Auto);
+			using var list = new PooledList<int>(data, ClearMode.Auto);
 			var dataCount = TestObjectCount / 10 > 0 ? TestObjectCount / 10 : TestObjectCount;
 			for (var i = data.Length - 1; i > data.Length - dataCount - 1; i--)
 			{
@@ -29,7 +29,7 @@ namespace Recyclable.Collections.Benchmarks
 		public void RecyclableList_Remove_LastItems()
 		{
 			var data = TestObjects;
-			using var list = new RecyclableList<long>(data);
+			using var list = new RecyclableList<int>(data);
 			var dataCount = TestObjectCount / 10 > 0 ? TestObjectCount / 10 : TestObjectCount;
 			for (var i = data.Length - 1; i > data.Length - dataCount - 1; i--)
 			{
@@ -40,7 +40,7 @@ namespace Recyclable.Collections.Benchmarks
 		public void RecyclableLongList_Remove_LastItems()
 		{
 			var data = TestObjects;
-			using var list = new RecyclableLongList<long>(data, minBlockSize: BlockSize, initialCapacity:base.TestObjectCount);
+			using var list = new RecyclableLongList<int>(data, minBlockSize: BlockSize, initialCapacity:base.TestObjectCount);
 			var dataCount = TestObjectCount / 10 > 0 ? TestObjectCount / 10 : TestObjectCount;
 			for (var i = data.Length - 1; i > data.Length - dataCount - 1; i--)
 			{
